requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
Click==7.0
importlib-metadata==1.3.0
more-itertools==8.0.2
packaging==19.2
pathspec==0.7.0
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
pyparsing==2.4.6
pytest==5.3.2
regex==2019.12.20
six==1.13.0
toml==0.10.0
typed-ast==1.4.0
wcwidth==0.1.7
zipp==0.6.0
